refactor(app): labware and module accessors
