+++ app/src/organisms/ModuleCard/index.tsx
@@ -46,7 +46,7 @@ import { UpdateBanner } from '/app/molecules/UpdateBanner'
 import { useChainLiveCommands } from '/app/resources/runs'
 import { useCurrentRunStatus } from '/app/organisms/RunTimeControl/hooks'
 import { useIsFlex } from '/app/redux-resources/robots'
-import { getModuleTooHot } from '/app/organisms/Devices/getModuleTooHot'
+import { getModuleTooHot } from '/app/transformations/modules'
 import { useToaster } from '/app/organisms/ToasterOven'
 import { MagneticModuleData } from './MagneticModuleData'
 import { TemperatureModuleData } from './TemperatureModuleData'
