+++ app/src/transformations/analysis/__tests__/getLabwareOffsetLocation.test.tsx
@@ -7,8 +7,10 @@ import {
   opentrons96PcrAdapterV1,
 } from '@opentrons/shared-data'
 import { getLabwareOffsetLocation } from '../getLabwareOffsetLocation'
-import { getLabwareLocation } from '../getLabwareLocation'
-import { getModuleInitialLoadInfo } from '/app/transformations/commands'
+import {
+  getModuleInitialLoadInfo,
+  getLabwareLocation,
+} from '/app/transformations/commands'
 import type {
   LoadedLabware,
   LoadedModule,
@@ -16,7 +18,6 @@ import type {
   CompletedProtocolAnalysis,
 } from '@opentrons/shared-data'
 
-vi.mock('../getLabwareLocation')
 vi.mock('/app/transformations/commands')
 
 const protocolWithTC = (multiple_tipacks_with_tc as unknown) as CompletedProtocolAnalysis
