+++ app/src/transformations/analysis/index.ts
@@ -1,2 +1,3 @@
 export * from './parseProtocolAnalysisOutput'
 export * from './getProtocolModulesInfo'
+export * from './getLabwareOffsetLocation'
