+++ app/src/transformations/commands/transformations/index.ts
@@ -4,3 +4,4 @@ export * from './formatDuration'
 export * from './formatInterval'
 export * from './getModuleInitialLoadInfo'
 export * from './getNestedLabwareInfo'
+export * from './getLabwareLocation'
